pyproject.toml
[tool.poetry]
name = "dex-gainers-losers"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "dex_gainers_losers"}]

[tool.poetry.dependencies]
python = "^3.9"
websockets = "^11.0.3"
poetry-dotenv-plugin = "^0.2.0"
python-telegram-bot = "^20.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

